#2 refactor Rental with adding dto

+++ src/main/java/com/api/chatop/service/RentalService.java
@@ -1,5 +1,7 @@
 package com.api.chatop.service;
 
+import com.api.chatop.dto.RequestPostRentalDto;
+import com.api.chatop.dto.RequestPutRentalDto;
 import com.api.chatop.model.Rental;
 import com.api.chatop.model.User;
 import com.api.chatop.repository.MessageRepository;
@@ -13,12 +15,10 @@ import org.springframework.web.multipart.MultipartFile;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
-import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-import java.util.Map;
 
 @Service
 @Transactional
@@ -32,53 +32,55 @@ public class RentalService {
 
     @Autowired
     private S3FileService s3FileService;
+    private final OffsetDateTime offsetDateTime = OffsetDateTime.now();
 
-    OffsetDateTime offsetDateTime = OffsetDateTime.now();
-
-    /* Update or save rental */
-    public void createRental(MultipartFile file, Map<String, String> rentalData, User user) {
-        String pictureUrl = getUrlFile(file);
-        Rental rental = new Rental();
+    /* create or save rental */
+    public void createRental(RequestPostRentalDto requestPostRentalDto, User user) {
+        String pictureUrl = getUrlFile(requestPostRentalDto.getPicture());
         if (pictureUrl != null) {
+            Rental rental = new Rental();
+            rental.setName(requestPostRentalDto.getName());
+            rental.setSurface(requestPostRentalDto.getSurface());
+            rental.setPrice(requestPostRentalDto.getPrice());
+            rental.setDescription(requestPostRentalDto.getDescription());
             rental.setPicture(pictureUrl);
             rental.setOwner(user);
             rental.setCreated_at(offsetDateTime);
             rental.setUpdated_at(offsetDateTime);
-            rental.setName(rentalData.get("name"));
-            rental.setSurface(new BigDecimal(rentalData.get("surface")));
-            rental.setPrice(new BigDecimal(rentalData.get("price")));
-            rental.setDescription(rentalData.get("description"));
-
             rentalRepository.save(rental);
         }
     }
 
     /* Update or save rental */
-    public void updateRental(Rental rental, User user) {
-        boolean isRentalPresent = rentalRepository.findById(rental.getId()).isPresent();
-        if (isRentalPresent) {
-            Rental rentalOldData = rentalRepository.findById(rental.getId()).get();
-            rental.setPicture(rentalOldData.getPicture());
-            rental.setOwner(user);
+    public void updateRental(RequestPutRentalDto requestPutRentalDto) {
+        if (getRental(requestPutRentalDto.getId()) != null) {
+            Rental rental = getRental(requestPutRentalDto.getId());
+            rental.setName(requestPutRentalDto.getName());
+            rental.setPrice(requestPutRentalDto.getPrice());
+            rental.setSurface(requestPutRentalDto.getSurface());
+            rental.setDescription(requestPutRentalDto.getDescription());
             rental.setUpdated_at(offsetDateTime);
-            rental.setCreated_at(rentalOldData.getCreated_at());
             rentalRepository.save(rental);
         }
     }
 
-    public void deleteRental(Integer id, User user) {
+    /* delete rental */
+    public String deleteRental(Integer id, User user) {
         Rental rental = getRental(id);
-        String url = rental.getPicture();
-        deleteFile(url.substring(url.lastIndexOf('/') + 1));
-        messageRepository.deleteByRentalId(id);
-        rentalRepository.deleteById(id);
+        if (rental.getOwner_id().equals(user.getId())) {
+            String url = rental.getPicture();
+            deleteFile(url.substring(url.lastIndexOf('/') + 1));
+            messageRepository.deleteByRentalId(id);
+            rentalRepository.deleteById(id);
+            return "Rental and related Messages are deleted";
+        }
+        return "Not authorized to delete";
     }
 
     /* get all rental */
     public List<Rental> getAllRentals(){
         List<Rental> rentals = new ArrayList<Rental>();
         rentalRepository.findAll().forEach(rental -> rentals.add(rental));
-
         return rentals;
     }
 
@@ -91,12 +93,13 @@ public class RentalService {
     }
 
 
-    public String deleteFile(final String fileName) {
+    /* delete file on S3 */
+    private String deleteFile(final String fileName) {
         s3FileService.deleteObject(s3FileService.getS3BucketName(), fileName);
         return "Deleted File: " + fileName;
     }
 
-
+    /* put file on S3 bucket */
     private File convertMultiPartToFile(MultipartFile file ) throws IOException {
         File convFile = new File(file.getOriginalFilename());
         FileOutputStream fos = new FileOutputStream(convFile);
@@ -105,7 +108,7 @@ public class RentalService {
         return convFile;
     }
 
-        /* check if valide url or upload file to s3 */
+    /* upload file to S3 */
     private String getUrlFile(MultipartFile multipartFile) {
         try {
             File file = convertMultiPartToFile(multipartFile);
@@ -125,7 +128,7 @@ public class RentalService {
         return null;
     }
 
-    /* check if valide extension  */
+    /* check if valid extension  */
     private Boolean isValideFile(String fileName) {
         List<String> allowedFileExtensions = new ArrayList<>(Arrays.asList("png", "jpg", "jpeg"));
         return allowedFileExtensions.contains(FilenameUtils.getExtension(fileName));
